improvement(save): don't prompt language selector on save
In v1 there was a setting to remember the current language. When saving and this setting was enabled
you would not get prompted to select a language. This change acts similar to that, except it never
prompts you to select a language when you save. Instead, it uses the default language you have
selected, which defaults to Text.